refactor(scan): centralize `getMarkStatus`
