+++ apps/central-scan/backend/src/types.ts
@@ -1,11 +1,4 @@
-import {
-  BallotLocale,
-  BallotMark,
-  BallotTargetMark,
-  MarkStatus,
-  MarkThresholds,
-  PageInterpretation,
-} from '@votingworks/types';
+import { BallotLocale, PageInterpretation } from '@votingworks/types';
 import { BallotStyleData } from '@votingworks/utils';
 
 export interface PageInterpretationWithAdjudication<
@@ -24,26 +17,4 @@ export interface BallotConfig extends BallotStyleData {
   filename: string;
   locales: BallotLocale;
   isLiveMode: boolean;
-}
-
-export function getMarkStatus(
-  mark: BallotTargetMark,
-  markThresholds: MarkThresholds
-): MarkStatus {
-  if (mark.score >= markThresholds.definite) {
-    return MarkStatus.Marked;
-  }
-
-  if (mark.score >= markThresholds.marginal) {
-    return MarkStatus.Marginal;
-  }
-
-  return MarkStatus.Unmarked;
-}
-
-export function isMarginalMark(
-  mark: BallotMark,
-  markThresholds: MarkThresholds
-): boolean {
-  return getMarkStatus(mark, markThresholds) === MarkStatus.Marginal;
 }
