+++ apps/scan/backend/src/types.ts
@@ -1,8 +1,6 @@
 import {
   AdjudicationReasonInfo,
-  BallotTargetMark,
   ElectionDefinition,
-  MarkStatus,
   MarkThresholds,
   PageInterpretation,
   PollsState,
@@ -29,21 +27,6 @@ export interface PageInterpretationWithAdjudication<
 export interface BallotPageQrcode {
   data: Uint8Array;
   position: 'top' | 'bottom';
-}
-
-export function getMarkStatus(
-  mark: BallotTargetMark,
-  markThresholds: MarkThresholds
-): MarkStatus {
-  if (mark.score >= markThresholds.definite) {
-    return MarkStatus.Marked;
-  }
-
-  if (mark.score >= markThresholds.marginal) {
-    return MarkStatus.Marginal;
-  }
-
-  return MarkStatus.Unmarked;
 }
 
 export type PrecinctScannerState =
