+++ libs/backend/src/scan/cast_vote_records/build_cast_vote_record.ts
@@ -16,12 +16,14 @@ import {
   getContests,
   InterpretedBmdPage,
   InterpretedHmpbPage,
+  MarkStatus,
   safeParseInt,
   SheetOf,
   VotesDict,
   YesNoContest,
   YesNoVote,
 } from '@votingworks/types';
+import { getMarkStatus } from '@votingworks/utils';
 
 import {
   BallotPageLayoutsLookup,
@@ -372,7 +374,10 @@ function buildOriginalSnapshot({
                 (Math.floor(mark.score * 100) / 100).toString(),
               ],
               HasIndication:
-                mark.score >= definiteMarkThreshold
+                getMarkStatus(mark.score, {
+                  marginal: definiteMarkThreshold,
+                  definite: definiteMarkThreshold,
+                }) === MarkStatus.Marked
                   ? CVR.IndicationStatus.Yes
                   : CVR.IndicationStatus.No,
             },
